Remove reference to getting started part 2
I've looked through the documentation and couldn't find a part two in order to fix this link to point to that. I'm assuming that there is no point 2, so it would be beneficial to remove the reference, but if there's somewhere else this should point, I'll update the PR.

Thanks!