improved grammar, expression
It's arguably better not to end a sentence with the preposition "for," even if the old rule "Never end a sentence with a preposition" isn't always true.

"Resetted" is not a word in English, so far as I know.  "Reset" is the past passive indicative form which should be used here.